Enable JSON output for Shapes (maps popup)
* add routes for RUIAN `kod` indentifers:
  * /shapes/by_parcel_stavobj/:parcel_id
  * /shapes/by_parcel_stavobj/:parcel_id/:stavebni_objekt_id
* add JSON formatter
* add posibility to show maps popup on landing map page
  by param `?openPopup=true`. It shows popup on the object
  in the center of map.

WARNING: Currently all events are returned evet to CORS requests.
It may be cachnged in the future (only a few first could be returned).

For all external apps/web pages using CORS requests it is necessary to
link to the mapasamospravy page for each event (link needs to be
fullfilled with events coordinates, e.g.
`http://mapasamospravy.cz/?openPopup=true#18/49.19920/16.58763`)